refactoring packages & added package 'entity'
