+++ src/main/java/ru/javarush/Main.java
@@ -1,4 +1,4 @@
-package ru.javarush.krivko;
+package ru.javarush;
 
 public class Main {
     public static void main(String[] args) {
