+++ src/main/resources/spy.properties
@@ -1,0 +1,4 @@
+driverlist=com.mysql.cj.jdbc.Driver
+dateformat=yyyy-MM-dd hh:mm:ss a
+appender=com.p6spy.engine.spy.appender.StdoutLogger
+logMessageFormat=com.p6spy.engine.spy.appender.MultiLineFormat
