+++ src/test/java/ru/javarush/MainTest.java
@@ -1,4 +1,4 @@
-package ru.javarush.krivko;
+package ru.javarush;
 
 public class MainTest {
